page config repository implementation
