+++ server/src/inversify.config.ts
@@ -14,8 +14,9 @@ import { EmailService } from "@/modules/email";
 import { MailjetEmailService } from "@/modules/email/MailjetEmailService";
 import { AuthProvider } from "@/modules/auth";
 import { FirebaseAuthProvider } from "@/modules/auth/FirebaseAuthProvider";
-import { PageRepository } from "./modules/pages/db";
-import { PageMongoRepository } from "./modules/pages/db/PageMongoRepository";
+import { PageRepository, PageConfigRepository } from "./modules/pages/db";
+import { PageMongoRepository } from "./modules/pages/db/page/PageMongoRepository";
+import { PageConfigMongoRepository } from "./modules/pages/db/page-config/PageConfigMongoRepository";
 import { PageService } from "./modules/pages";
 import { PageServiceImpl } from "./modules/pages/PageServiceImpl";
 import { ImageService } from "./modules/image";
@@ -29,6 +30,7 @@ container.bind(AuthProvider).to(FirebaseAuthProvider);
 container.bind(ProfileRepository).to(ProfileMongoRepository);
 container.bind(CommentRepository).to(CommentMongoRepository);
 container.bind(PageRepository).to(PageMongoRepository);
+container.bind(PageConfigRepository).to(PageConfigMongoRepository);
 
 container.bind(UrlService).to(UrlServiceImpl);
 container.bind(ImageService).to(ImageServiceImpl);
