+++ server/src/modules/pages/db/entity/PageConfigEntity.ts
@@ -1,0 +1,4 @@
+export interface PageConfigEntity {
+  id: string;
+  preserveQueryParams: string[];
+}
