+++ server/src/modules/pages/index.ts
@@ -3,7 +3,22 @@ import { UserId } from "@/modules/auth";
 
 export type { Page, PageDetails };
 
+export interface PageData {
+  websiteId: string;
+  pageId: string;
+  url: {
+    raw: string;
+    normalized: string;
+  };
+  meta: {
+    logo: string | null;
+    title: string | null;
+  };
+}
+
 export abstract class PageService {
+  abstract getPageData(url: string): Promise<PageData>;
+
   abstract getTopCommentedPages(
     commentsMinVoteSum: number,
     limit: number,
